Add CDC module for passing consonantly change bus
Continuous transfer of bus data using a variation of the 1-deep, 2-register
fifo. Operates in both slow2fast or fast2slow directions but has no
guaranties that all data will be passed. Can guaranty that the data that is
passed over is correct and in order.

For more information about the techniques employed, look in the following
references.

http://www.sunburst-design.com/papers/CummingsSNUG2008Boston_CDC.pdf (p. 55)
http://www.sunburst-design.com/papers/CummingsSNUG2008Boston_CDC.pdf (p. 49)